ml_service/requirements.txt
pandas==1.5.3
numpy==1.24.2
scikit-learn==1.2.1
kafka-python==2.0.1